requirements.txt
Flask==3.0.0
azure-ai-textanalytics==5.3.0
azure-core==1.29.5
python-dotenv==1.0.0
msal==1.24.0
requests==2.31.0
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
aiohttp==3.9.0
